fine-tune/requirements.txt
# PyTorch with CUDA 12.4 — install separately first:
# pip install torch==2.6.0 torchvision==0.21.0 --index-url https://download.pytorch.org/whl/cu124

# Core fine-tuning stack
transformers>=4.45.0
peft>=0.14.0
bitsandbytes>=0.45.0
accelerate>=0.34.0
trl>=0.12.0

# Qwen2-VL image utilities
qwen-vl-utils>=0.0.8

# Data & images
datasets>=3.0.0
Pillow>=10.0.0

# Optional: Anthropic API for auto-generating training labels
anthropic>=0.40.0
